@@ -1,6 +1,6 @@
 {
   "name": "fool-translations",
-  "version": "0.0.2",
+  "version": "0.1.0",
   "private": true,
   "type": "module",
   "scripts": {
@@ -11,7 +11,8 @@
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./jsconfig.json --watch",
     "test": "vitest",
     "lint": "prettier --check . && eslint .",
-    "format": "prettier --write ."
+    "format": "prettier --write .",
+    "cf-typegen": "wrangler types ./src/worker-configuration.d.ts"
   },
   "dependencies": {
     "@tiptap/core": "^2.8.0",
@@ -21,42 +22,42 @@
     "arctic": "^3.6.0",
     "clsx": "^2.1.1",
     "dayjs": "^1.11.13",
-    "devalue": "^5.1.1",
+    "drizzle-orm": "^0.44.2",
     "embla-carousel-autoplay": "^8.3.1",
-    "mode-watcher": "^0.4.1",
-    "set-cookie-parser": "^2.7.1",
+    "mode-watcher": "^1.0.7",
     "svelte-inview": "^4.0.4",
     "svelte-sonner": "^0.3.28",
-    "tailwind-merge": "^2.5.3",
+    "tailwind-merge": "^2.6.0",
     "tailwind-variants": "^0.2.1"
   },
   "devDependencies": {
     "@internationalized/date": "^3.8.0",
-    "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "@sveltejs/adapter-cloudflare": "^7.0.3",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@tailwindcss/typography": "^0.5.16",
     "@types/eslint": "^9.6.0",
-    "@types/set-cookie-parser": "^2.4.10",
     "autoprefixer": "^10.4.21",
     "bits-ui": "^0.22.0",
     "cmdk-sv": "^0.0.18",
+    "drizzle-kit": "^0.31.1",
     "embla-carousel-svelte": "^8.3.1",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.36.0",
+    "eslint-plugin-svelte": "^2.45.1",
     "formsnap": "^1.0.1",
     "globals": "^15.0.0",
     "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
+    "prettier-plugin-svelte": "^3.2.6",
     "prettier-plugin-tailwindcss": "^0.6.5",
-    "svelte": "^4.2.7",
+    "svelte": "^5.0.0",
     "svelte-check": "^4.0.0",
-    "sveltekit-superforms": "^2.19.1",
+    "sveltekit-superforms": "^2.26.1",
     "tailwindcss": "^3.4.9",
-    "typescript": "^5.0.0",
-    "vite": "^5.0.3",
+    "typescript": "^5.5.0",
+    "vite": "^5.4.19",
     "vitest": "^2.0.0",
-    "zod": "^3.23.8"
+    "wrangler": "^4.19.1",
+    "zod": "^3.25.56"
   }
 }
