@@ -1,47 +1,47 @@
 <script lang="ts">
   import Card from '$lib/components/novels/card.svelte';
+  import Seo from '$lib/components/SEO.svelte';
   import Input from '$lib/components/ui/input/input.svelte';
   import Label from '$lib/components/ui/label/label.svelte';
   import * as Tabs from '$lib/components/ui/tabs';
 
-  export let data;
+  let { data } = $props();
 
-  const { novels } = data;
-  let sort: 'popular' | 'recent' | 'views' = 'popular';
-  let status: 'all' | 'ongoing' | 'completed' = 'all';
+  let sort = $state<'popular' | 'recent' | 'views'>('popular');
+  let status = $state<'all' | 'ongoing' | 'completed'>('all');
+  let searchTitle = $state('');
 
-  let searchTitle: string = '';
-  $: sortedNovels = novels
-    .filter((novel) => {
-      if (status === 'all') return true;
-      return novel.status === status;
-    })
-    .sort((a, b) => {
-      if (sort === 'popular') {
-        return b.popularityScore - a.popularityScore;
-      } else if (sort === 'views') {
-        return b.totalViews - a.totalViews;
-      } else if (sort === 'recent') {
-        return new Date(b.lastUpdated).getTime() - new Date(a.lastUpdated).getTime();
-      }
-      return 0;
-    })
-    .filter((novel) => novel.title.toLowerCase().includes(searchTitle));
+  let sortedNovels = $derived(
+    data.novels
+      .filter((novel) => {
+        if (status === 'all') return true;
+        return novel.status === status;
+      })
+      .sort((a, b) => {
+        if (sort === 'popular') {
+          return b.popularityScore - a.popularityScore;
+        } else if (sort === 'views') {
+          return b.totalViews - a.totalViews;
+        } else if (sort === 'recent') {
+          return new Date(b.lastUpdated).getTime() - new Date(a.lastUpdated).getTime();
+        }
+        return 0;
+      })
+      .filter((novel) => novel.title.toLowerCase().includes(searchTitle))
+  );
 </script>
 
-<svelte:head>
-  <title>Fool Translations | Explore Popular Novels</title>
-</svelte:head>
+<Seo title="Explore Popular Novels | Fool Translations" />
 
 <div class="container mt-14 overflow-hidden md:mt-0">
   <div class="my-2 md:mx-20">
     <div class="mx-auto flex h-12 w-full flex-row items-center gap-2">
-      <Input placeholder="Search" bind:value={searchTitle} class="border-border" />
+      <Input placeholder="Search" bind:value={searchTitle} class="border-border" name="title" />
     </div>
     <div class="flex w-full flex-row gap-4">
       <div class="items-center gap-4 sm:flex sm:flex-row">
         <div class="flex flex-col gap-y-1 sm:w-auto">
-          <Label class="capitalize">{sort}</Label>
+          <p class="py-1 capitalize">{sort}</p>
           <Tabs.Root bind:value={sort} id="sort">
             <Tabs.List class="gap-2">
               <Tabs.Trigger value="popular">Popular</Tabs.Trigger>
@@ -53,7 +53,7 @@
       </div>
       <div class="items-center gap-4 sm:flex sm:flex-row">
         <div class="flex flex-col gap-y-1 sm:w-auto">
-          <Label class="capitalize">{status}</Label>
+          <p class="py-1 capitalize">{status}</p>
           <Tabs.Root bind:value={status} id="status">
             <Tabs.List class="gap-2">
               <Tabs.Trigger value="all">All</Tabs.Trigger>
@@ -64,9 +64,6 @@
         </div>
       </div>
     </div>
-
-    <!-- <pre>{JSON.stringify(sortedNovels, null, 4)}</pre> -->
-
     <div class="h my-4 grid grid-cols-2 gap-4 md:grid-cols-4">
       {#each sortedNovels as novel}
         <Card id={novel.id} image={novel.cover} title={novel.title} />
