@@ -1,18 +1,19 @@
 // See https://kit.svelte.dev/docs/types#app
-
-import type { Session, User } from '$lib/types';
-
 // for information about these interfaces
 declare global {
   namespace App {
     // interface Error {}
     interface Locals {
-      user: User | null;
-      session: Session | null;
+      user: import('$lib/server/auth').SessionValidationResult['user'];
+      session: import('$lib/server/auth').SessionValidationResult['session'];
     }
     // interface PageData {}
     // interface PageState {}
-    // interface Platform {}
+    interface Platform {
+      env: Env;
+      cf: CfProperties;
+      ctx: ExecutionContext;
+    }
   }
   interface Window {
     turnstile: {
