@@ -1,24 +1,29 @@
 import {
-  deleteSessionTokenCookie,
+  validateSessionToken,
   setSessionTokenCookie,
-  validateSessionToken
-} from '$lib/session';
+  deleteSessionTokenCookie
+} from '$lib/server/auth';
 import type { Handle } from '@sveltejs/kit';
 
 export const handle: Handle = async ({ event, resolve }) => {
   const token = event.cookies.get('session') ?? null;
+
+  // if (event.url.pathname.startsWith('/.well-known/appspecific/com.chrome.devtools')) {
+  //   return new Response(null, { status: 204 }); // Return empty response with 204 No Content
+  // }
+
   if (token === null) {
     event.locals.user = null;
     event.locals.session = null;
     return resolve(event);
   }
 
-  const { session, user } = await validateSessionToken(event);
-  if (!session) {
+  const { session, user } = await validateSessionToken(event.platform!.env.DB, token);
+  if (session !== null) {
+    setSessionTokenCookie(event, token, session.expiresAt);
+  } else {
     deleteSessionTokenCookie(event);
-    return resolve(event);
   }
-  setSessionTokenCookie(event, token, session!.expiresAt);
 
   event.locals.session = session;
   event.locals.user = user;
