@@ -1,12 +1,17 @@
 import type { PageServerLoad } from './$types';
-import { BE_URL } from '$env/static/private';
-import type { ApiResponse, Notification } from '$lib/types';
+import { createDB } from '$lib/server/db';
+import { notification } from '$lib/server/db/schema/notification.schema';
+import { and, eq } from 'drizzle-orm';
 
-export const load: PageServerLoad = async ({ fetch, locals }) => {
+export const load: PageServerLoad = async ({ locals, platform }) => {
   if (locals.user) {
-    const res = await fetch(`${BE_URL}/notifications`);
-    const data: ApiResponse<Notification[]> = await res.json();
+    const notifications = await createDB(platform!.env.DB).query.notification.findMany({
+      where: and(eq(notification.userId, locals.user.id), eq(notification.type, 'new_chapter')),
+      columns: { novelId: false },
+      with: { novel: { columns: { id: true, cover: true, title: true } } },
+      orderBy: (notification, { desc }) => [desc(notification.createdAt)]
+    });
 
-    return { chapters: data.data };
+    return { chapters: notifications };
   }
 };
