@@ -1,3 +1,4 @@
+import type { Chapter } from '$lib/server/db/schema/novel.schema.js';
 import { error } from '@sveltejs/kit';
 
 export async function load({ data, fetch }) {
@@ -5,7 +6,7 @@ export async function load({ data, fetch }) {
   if (!res.ok) throw error(404, 'Chapter not found');
 
   const mockChapters = await fetch('/MOCK_DATA_chapters.json');
-  const chapters = await mockChapters.json();
+  const chapters: Chapter[] = await mockChapters.json();
   const content: string = await res.text();
   return { ...data, content, chapters };
 }
