@@ -1,54 +1,52 @@
-import { error, type Actions } from '@sveltejs/kit';
+import type { Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
-import { fail, superValidate } from 'sveltekit-superforms';
-import { zod } from 'sveltekit-superforms/adapters';
-import { chapterSchema } from '../schema';
-import { BE_URL, CLOUDINARY_CLOUD_NAME, CLOUDINARY_CLOUD_PRESET } from '$env/static/private';
-import type { CloudinaryResponse } from '$lib/types';
+import { fail, superValidate, message } from 'sveltekit-superforms';
+import { zod4 } from 'sveltekit-superforms/adapters';
+import { chapterSchema } from '../admin.validation';
+import {
+  addChapter,
+  getNovel,
+  notifySubscribers,
+  updateLastUpdatedForNovel
+} from '$lib/server/service';
+import { requireAdmin, UNIQUE_CONSTRAINT } from '$lib/server/utils';
 
 export const load: PageServerLoad = async ({ locals }) => {
-  if (!locals.user || locals.user.type !== 'admin') {
-    return error(404, 'Page Not Found');
-  }
-
-  return { form: await superValidate(zod(chapterSchema)) };
+  requireAdmin(locals);
+  return { form: await superValidate(zod4(chapterSchema)) };
 };
 
-export const actions: Actions = {
+export const actions = {
   create: async (event) => {
-    const form = await superValidate(event, zod(chapterSchema));
+    requireAdmin(event.locals);
+
+    const form = await superValidate(event, zod4(chapterSchema));
     if (!form.valid) return fail(400, { form });
 
-    const blob = new Blob([form.data.content], { type: 'text/markdown' });
-    const fileName = `${form.data.novelId.slice(4)}-chapter-${form.data.chapterNum}.md`;
+    const currentNovel = await getNovel(event.platform!.env, form.data.novelId, false);
+    if (!currentNovel) return message(form, 'Novel Not Found', { status: 404 });
 
-    const formData = new FormData();
-    formData.append('file', blob, fileName);
-    formData.append('upload_preset', CLOUDINARY_CLOUD_PRESET);
-    formData.append('resource_type', 'raw');
-
-    const uploadChapter = await fetch(
-      `https://api.cloudinary.com/v1_1/${CLOUDINARY_CLOUD_NAME}/raw/upload`,
-      { method: 'POST', body: formData }
-    );
-
-    const chapterData: CloudinaryResponse = await uploadChapter.json();
-
-    const res = await fetch(`${BE_URL}/admin/chapter`, {
-      method: 'POST',
-      body: JSON.stringify({
+    try {
+      const newChapter = await addChapter(event.platform!.env, {
         title: form.data.title,
-        novelId: form.data.novelId,
-        chapterNum: Number(form.data.chapterNum),
-        content: chapterData.secure_url
-      }),
-      headers: {
-        'Content-type': 'application/json',
-        Cookie: `session=${event.cookies.get('session')}`
-      }
-    });
-    const newChapter = await res.json();
+        novelId: currentNovel.id,
+        chapterNum: form.data.chapterNum,
+        content: form.data.content
+      });
 
-    return { form, newChapter };
+      if (newChapter) {
+        updateLastUpdatedForNovel(event.platform!.env, newChapter.novelId);
+        notifySubscribers(event.platform!.env, newChapter.novelId);
+      }
+
+      return { form, chapter: newChapter };
+    } catch (err: any) {
+      if (String(err.cause).includes(UNIQUE_CONSTRAINT)) {
+        return message(form, `Chapter ${form.data.chapterNum} Already Exists`, { status: 409 });
+      }
+
+      console.error(err);
+      return message(form, 'Internal Server Errror', { status: 500 });
+    }
   }
-};
+} satisfies Actions;
