@@ -1,38 +1,34 @@
-import { error, type Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
-import { superValidate } from 'sveltekit-superforms';
-import { zod } from 'sveltekit-superforms/adapters';
-import { updateNovelSchema } from '../../schema';
-import { fail } from '@sveltejs/kit';
-import { BE_URL } from '$env/static/private';
+import { message, superValidate } from 'sveltekit-superforms';
+import { zod4 } from 'sveltekit-superforms/adapters';
+import { updateNovelSchema } from '../../admin.validation';
+import { fail, type Actions } from '@sveltejs/kit';
+import { requireAdmin } from '$lib/server/utils';
+import { updateNovel } from '$lib/server/service';
 
 export const load: PageServerLoad = async ({ locals }) => {
-  if (!locals.user || locals.user.type !== 'admin') {
-    return error(404, 'Page Not Found');
-  }
-
-  return { form: await superValidate(zod(updateNovelSchema)) };
+  requireAdmin(locals);
+  return { form: await superValidate(zod4(updateNovelSchema)) };
 };
 
-export const actions: Actions = {
+export const actions = {
   default: async (event) => {
-    const form = await superValidate(event, zod(updateNovelSchema));
+    requireAdmin(event.locals);
+
+    const form = await superValidate(event, zod4(updateNovelSchema));
     if (!form.valid) return fail(400, { form });
 
     const arrGenres = form.data.genres.split(',');
     const bodyData = { ...form.data, genres: arrGenres };
 
-    const res = await fetch(`${BE_URL}/admin/novel/${event.url.searchParams.get('novelId')}`, {
-      method: 'PUT',
-      body: JSON.stringify(bodyData),
-      headers: {
-        'Content-type': 'application/json',
-        Cookie: `session=${event.cookies.get('session')}`
-      }
-    });
+    try {
+      const updatedNovel = await updateNovel(event.platform!.env, bodyData, form.data.id);
+      if (!updatedNovel) return message(form, 'Novel Not Found', { status: 404 });
 
-    const updatedNovel = await res.json();
-
-    return { form, updatedNovel };
+      return { form, novel: updatedNovel };
+    } catch (err) {
+      console.error(err);
+      return message(form, 'Internal Server Errror', { status: 500 });
+    }
   }
-};
+} satisfies Actions;
