@@ -1,25 +1,25 @@
 import { redirect, type Actions } from '@sveltejs/kit';
 import type { PageServerLoad } from './$types';
-import { fail, superValidate } from 'sveltekit-superforms';
-import { zod } from 'sveltekit-superforms/adapters';
-import { loginSchema, type AuthResponse } from '../schema';
-import { BE_URL, TURNSLITE_SECRET_KEY } from '$env/static/private';
-import { parseString } from 'set-cookie-parser';
-import { setSessionTokenCookie } from '$lib/session';
+import { fail, message, superValidate } from 'sveltekit-superforms';
+import { zod4 } from 'sveltekit-superforms/adapters';
+import { TURNSLITE_SECRET_KEY } from '$env/static/private';
 import type { TurnstileResponse } from '$lib/types';
 import { PUBLIC_APP_ENV } from '$env/static/public';
+import { createDB } from '$lib/server/db';
+import { and, eq } from 'drizzle-orm';
+import { user } from '$lib/server/db/schema/user.schema';
+import { createSession, generateSessionToken, setSessionTokenCookie } from '$lib/server/auth';
+import { loginSchema } from '../auth.validation';
 
 export const load: PageServerLoad = async ({ locals }) => {
-  if (locals.user) {
-    return redirect(302, '/');
-  }
+  if (locals.user) redirect(302, '/');
 
-  return { form: await superValidate(zod(loginSchema)) };
+  return { form: await superValidate(zod4(loginSchema)) };
 };
 
-export const actions: Actions = {
+export const actions = {
   default: async (event) => {
-    const form = await superValidate(event, zod(loginSchema));
+    const form = await superValidate(event, zod4(loginSchema));
     if (!form.valid) return fail(400, { form });
 
     const verify = await fetch('https://challenges.cloudflare.com/turnstile/v0/siteverify', {
@@ -35,21 +35,33 @@ export const actions: Actions = {
     });
 
     const turnstile: TurnstileResponse = await verify.json();
-    if (turnstile.success) {
-      const res = await fetch(`${BE_URL}/auth/signin`, {
-        method: 'POST',
-        body: JSON.stringify(form.data),
-        headers: { 'Content-type': 'application/json' }
+    if (!turnstile.success) {
+      return message(form, `Turnstile verification failed: ${turnstile?.['error-codes'] || ''}`, {
+        status: 401
       });
-      const signin: AuthResponse = await res.json();
-      if (signin.success) {
-        const cookie = parseString(res.headers.get('set-cookie')!);
-        setSessionTokenCookie(event, cookie.value, cookie.expires as Date);
+    }
+
+    try {
+      const userExist = await createDB(event.platform!.env.DB).query.user.findFirst({
+        where: and(eq(user.email, form.data.email.toLowerCase()), eq(user.provider, 'credentials')),
+        columns: { id: true, password: true }
+      });
+      if (!userExist) {
+        return message(form, 'Invalid Email or Password.', { status: 401 });
       }
 
-      return { form, signin };
-    } else {
-      return { form, turnstile };
+      if (userExist.password !== form.data.password) {
+        return message(form, 'Invalid Email or Password', { status: 401 });
+      }
+
+      const token = generateSessionToken();
+      const session = await createSession(event.platform!.env.DB, token, userExist?.id);
+      setSessionTokenCookie(event, token, session.expiresAt);
+
+      return { form, success: true };
+    } catch (err) {
+      console.error(err);
+      return message(form, 'Internal Server Errror', { status: 500 });
     }
   }
-};
+} satisfies Actions;
