@@ -1,16 +1,17 @@
 import { google } from '$lib/oauth';
-
 import { error, type RequestEvent } from '@sveltejs/kit';
 import { ArcticFetchError, OAuth2RequestError } from 'arctic';
-import { BE_URL } from '$env/static/private';
-import { parseString } from 'set-cookie-parser';
-import { setSessionTokenCookie } from '$lib/session';
+import { createSession, generateSessionToken, setSessionTokenCookie } from '$lib/server/auth';
+import type { GoogleOAuthResponse } from '$lib/types';
+import { generateRandId } from '$lib/server/utils';
+import { createUser, getUserFromProviderId } from '$lib/server/service/auth.service';
 
 export async function GET(event: RequestEvent): Promise<Response> {
   const code = event.url.searchParams.get('code');
   const state = event.url.searchParams.get('state');
   const storedState = event.cookies.get('google_oauth_state') ?? null;
   const codeVerifier = event.cookies.get('google_code_verifier') ?? null;
+
   if (code === null || state === null || storedState === null || codeVerifier === null) {
     return new Response(null, { status: 400 });
   }
@@ -20,34 +21,46 @@ export async function GET(event: RequestEvent): Promise<Response> {
 
   try {
     let tokens = await google.validateAuthorizationCode(code, codeVerifier);
-
-    const res = await fetch(`${BE_URL}/oauth/google`, {
-      headers: { Cookie: `access_token=${tokens.accessToken()}` }
+    const response = await fetch('https://openidconnect.googleapis.com/v1/userinfo', {
+      headers: { Authorization: `Bearer ${tokens.accessToken()}` }
     });
-    const data: { success: boolean } = await res.json();
 
-    if (!data.success) throw error(500, 'Authentication failed');
+    const user: GoogleOAuthResponse = await response.json();
 
-    const cookie = parseString(res.headers.get('set-cookie')!);
-    setSessionTokenCookie(event, cookie.value, cookie.expires as Date);
+    const userExist = await getUserFromProviderId(event.platform!.env, user.sub);
+    if (userExist) {
+      const token = generateSessionToken();
+      const session = await createSession(event.platform!.env.DB, token, userExist?.id);
+      setSessionTokenCookie(event, token, session.expiresAt);
 
-    return new Response(null, {
-      status: 302,
-      headers: {
-        Location: '/'
-      }
+      return new Response(null, { status: 302, headers: { Location: '/' } });
+    }
+
+    const userId = generateRandId('usr');
+    await createUser({
+      env: event.platform!.env,
+      email: user.email.toLowerCase(),
+      id: userId,
+      name: user.name,
+      avatar: user.picture,
+      provider: 'google',
+      providerId: user.sub
     });
+
+    const token = generateSessionToken();
+    const session = await createSession(event.platform!.env.DB, token, userId);
+    setSessionTokenCookie(event, token, session.expiresAt);
+
+    return new Response(null, { status: 302, headers: { Location: '/' } });
   } catch (e) {
-    console.log(e);
-
     if (e instanceof OAuth2RequestError) {
-      throw error(500, e.message);
+      return error(500, e.message);
     }
     if (e instanceof ArcticFetchError) {
       const cause = e.cause;
-      throw error(500, cause as string);
+      return error(500, cause as string);
     }
-
-    throw error(500, 'Authentication failed');
+    console.log(e);
+    return error(500, 'Authentication failed');
   }
 }
